requirements.txt
future==0.18.2
protobuf==3.11.3
ruamel.yaml==0.16.10
jsonlines==1.2.0
ai2thor==2.3.10
progressbar
flask
torch
torchvision
